Add a flag to publish metrics with high resolution (#32)
So far, this tool is only publishing metrics with high resolution when
the original prometheus metrics have the label __cw_high_res.

However, in our case we're running exporters and prom-to-cloudwatch
as sidecars of ECS tasks. This tool being directly connected to the
exporters, we can't add the high res label.